feat: use vscode's implementation to open directory
Instead of relying on explorer being in the PATH environment variable, we rely on vscode to find the explorer for the specific os.
We open snippet directory by opening  /path/to/folder/placeholder,
since vscode opens the directory that the file/directory is in instead of opening the directory.
Hopefully vscode doesn't break this command.